+++ .github/workflows/ci.yml
@@ -33,8 +33,13 @@ jobs:
           java-version: 17
           distribution: 'temurin'
 
-      - name: Test with Maven
+      - name: Maven Test With Spring 6.x
         run: mvn --batch-mode test -Dsurefire.jdk-toolchain-version=${{ matrix.java }}
+        if: ${{ matrix.java >= 17 }}
+
+      - name: Maven Test Without Spring 6.x
+        run: mvn --batch-mode test -Dsurefire.jdk-toolchain-version=${{ matrix.java }} -Dskip.spring.v6x.test=true
+        if: ${{ matrix.java < 17 }}
 
       - name: Build with Maven
         run: mvn install -DskipTests=true -Dmaven.javadoc.skip=true -B -V -DminimumPriority=1
