+++ sentinel-adapter/pom.xml
@@ -27,8 +27,9 @@
         <module>sentinel-spring-webflux-adapter</module>
         <module>sentinel-api-gateway-adapter-common</module>
         <module>sentinel-spring-cloud-gateway-adapter</module>
+        <module>sentinel-web-adapter-common</module>
         <module>sentinel-spring-webmvc-adapter</module>
-        <module>sentinel-spring-webmvc-6x-adapter</module>
+        <module>sentinel-spring-webmvc-v6x-adapter</module>
         <module>sentinel-zuul2-adapter</module>
         <module>sentinel-okhttp-adapter</module>
         <module>sentinel-jax-rs-adapter</module>
