--- sentinel-adapter/sentinel-spring-webmvc-6x-adapter/src/main/java/com/alibaba/csp/sentinel/adapter/spring/webmvc/callback/DefaultBlockExceptionHandler.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 1999-2019 Alibaba Group Holding Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.alibaba.csp.sentinel.adapter.spring.webmvc.callback;
-
-import com.alibaba.csp.sentinel.slots.block.BlockException;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.PrintWriter;
-
-/**
- * Default handler for the blocked request.
- *
- * @author kaizi2009
- */
-public class DefaultBlockExceptionHandler implements BlockExceptionHandler {
-
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, BlockException e) throws Exception {
-        // Return 429 (Too Many Requests) by default.
-        response.setStatus(429);
-
-        PrintWriter out = response.getWriter();
-        out.print("Blocked by Sentinel (flow limiting)");
-        out.flush();
-        out.close();
-    }
-
-}
