--- sentinel-adapter/sentinel-spring-webmvc-6x-adapter/src/test/java/com/alibaba/csp/sentinel/adapter/spring/webmvc/SentinelWebInterceptorTest.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 1999-2018 Alibaba Group Holding Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.alibaba.csp.sentinel.adapter.spring.webmvc;
-
-import com.alibaba.csp.sentinel.adapter.spring.webmvc.config.SentinelWebMvcConfig;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * @author Eric Zhao
- */
-public class SentinelWebInterceptorTest {
-
-    @Test(expected = IllegalArgumentException.class)
-    public void testPassIllegalConfig() {
-        SentinelWebMvcConfig config = new SentinelWebMvcConfig();
-        config.setRequestAttributeName(null);
-        SentinelWebInterceptor interceptor = new SentinelWebInterceptor(config);
-    }
-}
