+++ sentinel-adapter/sentinel-spring-webmvc-v6x-adapter/src/main/java/com/alibaba/csp/sentinel/adapter/spring/webmvc_v6x/config/SentinelPreWebMvcConfig.java
@@ -1,0 +1,57 @@
+package com.alibaba.csp.sentinel.adapter.spring.webmvc_v6x.config;
+
+import com.alibaba.csp.sentinel.adapter.web.common.UrlCleaner;
+
+/**
+ * @since 1.8.8
+ */
+public class SentinelPreWebMvcConfig extends BaseWebMvcConfig {
+
+    public static final String DEFAULT_REQUEST_ATTRIBUTE_NAME = "$$sentinel_pre_spring_web_entry_attr";
+
+    private UrlCleaner urlCleaner;
+
+    /**
+     * Specify whether the URL resource name should contain the HTTP method prefix (e.g. {@code POST:}).
+     */
+    private boolean httpMethodSpecify;
+
+    /**
+     * Specify whether unify web context(i.e. use the default context name), and is true by default.
+     *
+     * @since 1.7.2
+     */
+    private boolean webContextUnify = true;
+
+    public SentinelPreWebMvcConfig() {
+        super();
+        setRequestAttributeName(DEFAULT_REQUEST_ATTRIBUTE_NAME);
+    }
+
+    public boolean isHttpMethodSpecify() {
+        return httpMethodSpecify;
+    }
+
+    public SentinelPreWebMvcConfig setHttpMethodSpecify(boolean httpMethodSpecify) {
+        this.httpMethodSpecify = httpMethodSpecify;
+        return this;
+    }
+
+    public boolean isWebContextUnify() {
+        return webContextUnify;
+    }
+
+    public SentinelPreWebMvcConfig setWebContextUnify(boolean webContextUnify) {
+        this.webContextUnify = webContextUnify;
+        return this;
+    }
+
+    public UrlCleaner getUrlCleaner() {
+        return urlCleaner;
+    }
+
+    public SentinelPreWebMvcConfig setUrlCleaner(UrlCleaner urlCleaner) {
+        this.urlCleaner = urlCleaner;
+        return this;
+    }
+}
